- fix: conversion EWarnings

+++ NaoTHSoccer/Source/Core/Cognition/Modules/Perception/VisualCortex/Detectors/Experimental/NeoLineDetector.cpp
@@ -630,8 +630,8 @@ void NeoLineDetector::calculateNeigbors(const std::vector<ScanLineEdgelPercept::
         // update neigbors
         if(s > threshold) 
         {
-          int idx_left = edgelTwo.id < edgelOne.id ? j : i;
-          int idx_right = edgelTwo.id < edgelOne.id ? i : j;
+          int idx_left = edgelTwo.id < edgelOne.id ? (int)j : (int)i;
+          int idx_right = edgelTwo.id < edgelOne.id ? (int)i : (int)j;
 
           Neighbors& left_node = neighbors[idx_left];
           Neighbors& right_node = neighbors[idx_right];
@@ -690,8 +690,8 @@ void NeoLineDetector::calculatePairsAndNeigbors(
         // update neigbors
         if(s > threshold) 
         {
-          int idx_left = edgelTwo.id < edgelOne.id ? j : i;
-          int idx_right = edgelTwo.id < edgelOne.id ? i : j;
+          int idx_left = edgelTwo.id < edgelOne.id ? (int)j : (int)i;
+          int idx_right = edgelTwo.id < edgelOne.id ? (int)i : (int)j;
 
           Neighbors& left_node = neighbors[idx_left];
           Neighbors& right_node = neighbors[idx_right];
